doc: git-rebase: update discussion of internals
The current discussion of git's internals in the DESCRIPTION section
describes the `apply` backend for `git rebase`. This updates it to
instead describe the `merge` backend, which is the default.

Also delete a duplicate explanation of how git rebase skips commits
with the same textual changes (it's explained in more detail a few
lines further down) and move the `ORIG_HEAD` not down so that it doesn't
interrupt the discussion of the mechanics.