doc: git-add: start man page with an example
This replaces the intro paragraph of the `git-add` man page with an
example to try to clarify it for new users. The goal here is use less
jargon but communicate essentially the same information.

Removed "The "index" holds a snapshot of the content..." from the beginning
since "the index" is a complex concept and I don't think the brief
explanation here is doing it justice.