doc: glossary: clarify the meaning of tree-ish
Right now it is not explained anywhere in the glossary that you can use
a branch name or other reference in any place that a Git command
specifies a `<tree-ish>` or <commit-ish>. Clarify this in the
documentation for `<tree-ish>`, since it's one of the most common ways
to refer to a <tree-ish> in practice.

Also mention the form `main:src/` for referencing a `tree-ish`, and
refer to `man gitrevisions` for more information.