push: make simple mode fail for branches with no upstream
With push.default=simple, `git push` was succeeding when pushing new
branches with no upstream configured to an upstream not named 'origin'.
For example `git push notorigin` would succeed when pushing a new branch.

The commit introducing `push.default=simple` (b55e677522) says that Git
should deny the push if there's no configured upstream, and as of that
commit `git push notorigin` would fail, as the commit message says it
should.

I ran a `git bisect` to figure out when `git push notorigin` started
succeeding in that situation, and found that it was in ed2b18292b.
The intent of that commit seems to be to fix the behaviour of
`push.default=simple` when an upstream is set, but it doesn't look like
it intended to change the behaviour when the upstream is _not_ set.

Always call get_upstream_ref() in simple mode, which will die with
an appropriate error message if the branch has no upstream branch
configured.

Add a test case which failed before this change and succeeds after.